refactor(dsp): move afe/graph spectrum into @wake-studio/dsp, rename kws decision logic
- afe/graph/core/dsp.ts -> packages/dsp/src/spectrum.ts: computeSpectrum,
  FFT_SIZE/SPECTRUM_BINS/HANN_WINDOW + sineWave/constant/argMax test
  helpers now live in the dsp package (ADR-032).
- afe/graph tests/dsp.test.ts -> packages/dsp/tests/behavior.test.ts:
  behavioral unit tests (downsample 3:1, levelDb thresholds, HANN
  symmetry, computeSpectrum peaks, ...) now run against the package.
- afe/graph spec l1 now points at packages/dsp tests; graph vitest gains
  passWithNoTests (scheduling is browser-e2e tested, no L1 remains).
- kws/engine/core/dsp.ts -> logic.ts: ScoreSmoother/TriggerDetector/VAD
  gate are decision logic, not DSP; renamed to match actual function.
- All module-level dsp.ts files are gone: graph deleted, rnnoise/few-shot
  renamed earlier, kws engine now logic.ts. @wake-studio/dsp is the sole
  DSP home.
- Full-repo typecheck/lint/unit green; web build bundles fft core into the
  worklet.
