+++ DECISIONS.md
@@ -804,9 +804,12 @@ Status legend: `Proposed` · `Accepted` · `Superseded` · `Deprecated`
   6. **Renamed non-DSP files (2026-08-06):** `afe/rnnoise/core/dsp.ts` ->
      `constants.ts` (RNNOISE_FRAME_SIZE, vadToProbability, dsp re-exports);
      `few-shot/core/dsp.ts` -> `quality.ts` (cosine similarity,
-     sample-quality checks). `afe/graph/core/dsp.ts` and
-     `kws/engine/core/dsp.ts` keep their names (they contain real
-     spectrum/DSP logic and decision logic respectively).
+     sample-quality checks); `kws/engine/core/dsp.ts` -> `logic.ts`
+     (ScoreSmoother/TriggerDetector/VAD gate - decision logic, not DSP).
+     `afe/graph/core/dsp.ts` was moved into `@wake-studio/dsp` as
+     `spectrum.ts` (computeSpectrum + test helpers) - the graph module now
+     has no in-module DSP file at all; its spec `l1` points at the dsp
+     package tests.
 - **Rationale:** (a) eliminates three duplicated hand-written FFTs and the
   PLiX/openWakeWord front-end drift risk; (b) the FFT core is a mature
   third-party implementation that can be fully audited; (c) higher-level
