+++ docs/modules/kws.md
@@ -386,7 +386,7 @@ All parameters are surfaced in the **Studio config panel** with the defaults bel
 
 - **Unit (Vitest):** the pure logic - score smoothing (sliding-window max),
   threshold + min-duration trigger logic, cooldown, VAD gate. These are extracted
-  to a testable module (like `afe/graph/dsp.ts`, itself delegating to
+  to a testable module (`kws/engine/core/logic.ts`, with numeric DSP in
   `@wake-studio/dsp` ADR-032) with no ONNX dependency.
 - **WASM runtime test (L2, Node):** load the sherpa-onnx-kws emscripten bundle in
   a Node process (the glue supports `ENVIRONMENT=node`), instantiate the
