+++ packages/modules/kws/engine/spec/module.spec.json
@@ -129,7 +129,7 @@
     }
   },
   "tests": {
-    "l1": "packages/modules/kws/engine/tests/dsp.test.ts",
+    "l1": "packages/modules/kws/engine/tests/logic.test.ts",
     "l2": "packages/modules/kws/engine/tests/backend.test.ts",
     "required": ["l1"]
   },
